Add quick sort algorithm; 1. add config, choose to use tim or quick 2. reformat: move clone, deserialize functions to abstract class IntTVList 3. fix the test
