+++ server/src/assembly/resources/conf/iotdb-datanode.properties
@@ -370,6 +370,10 @@ timestamp_precision=ms
 # Datatype: long
 # unseq_memtable_flush_check_interval_in_ms=600000
 
+# The sort algorithms used in the memtable's TVList
+# 0: default tim sort, 1: quick, 2: backward
+tvList_sort_algorithm=0
+
 # When the average point number of timeseries in memtable exceeds this, the memtable is flushed to disk. The default threshold is 100000.
 # Datatype: int
 # avg_series_point_number_threshold=100000
