+++ server/src/main/java/org/apache/iotdb/db/utils/datastructure/IntTVList.java
@@ -29,10 +29,13 @@ import org.apache.iotdb.tsfile.read.common.block.TsBlockBuilder;
 import org.apache.iotdb.tsfile.utils.BitMap;
 import org.apache.iotdb.tsfile.utils.TsPrimitiveType;
 
+import java.io.DataInputStream;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.apache.iotdb.db.rescon.PrimitiveArrayManager.ARRAY_SIZE;
+import static org.apache.iotdb.db.rescon.PrimitiveArrayManager.TVLIST_SORT_ALGORITHM;
 
 public abstract class IntTVList extends TVList {
   // list of primitive array, add 1 when expanded -> int primitive array
@@ -44,8 +47,16 @@ public abstract class IntTVList extends TVList {
     values = new ArrayList<>();
   }
 
-  public TimIntTVList clone() {
-    TimIntTVList cloneList = new TimIntTVList();
+  public static IntTVList newList() {
+    if (TVLIST_SORT_ALGORITHM == 1) {
+      return new QuickIntTVList();
+    }
+    return new TimIntTVList();
+  }
+
+  @Override
+  public IntTVList clone() {
+    IntTVList cloneList = IntTVList.newList();
     cloneAs(cloneList);
     for (int[] valueArray : values) {
       cloneList.values.add(cloneValue(valueArray));
@@ -236,5 +247,18 @@ public abstract class IntTVList extends TVList {
       buffer.putLong(getTime(rowIdx));
       buffer.putInt(getInt(rowIdx));
     }
+  }
+
+  public static IntTVList deserialize(DataInputStream stream) throws IOException {
+    IntTVList tvList = IntTVList.newList();
+    int rowCount = stream.readInt();
+    long[] times = new long[rowCount];
+    int[] values = new int[rowCount];
+    for (int rowIdx = 0; rowIdx < rowCount; ++rowIdx) {
+      times[rowIdx] = stream.readLong();
+      values[rowIdx] = stream.readInt();
+    }
+    tvList.putInts(times, values, null, 0, rowCount);
+    return tvList;
   }
 }
