+++ server/src/main/java/org/apache/iotdb/db/utils/datastructure/QuickAlignedTVList.java
@@ -4,42 +4,37 @@ import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 
 import java.util.List;
 
-public class QuickAlignedTVList extends AlignedTVList implements QuickSort{
-    QuickAlignedTVList(List<TSDataType> types) {
-        super(types);
-    }
+public class QuickAlignedTVList extends AlignedTVList implements QuickSort {
+  QuickAlignedTVList(List<TSDataType> types) {
+    super(types);
+  }
 
-    @Override
-    public void sort() {
-        qsort(0, rowCount - 1);
-    }
+  @Override
+  public void sort() {
+    qsort(0, rowCount - 1);
+  }
 
-    @Override
-    protected void set(int src, int dest) {
-        long srcT = getTime(src);
-        int srcV = getValueIndex(src);
-        set(dest, srcT, srcV);
-    }
+  @Override
+  protected void set(int src, int dest) {
+    long srcT = getTime(src);
+    int srcV = getValueIndex(src);
+    set(dest, srcT, srcV);
+  }
 
-    @Override
-    public AlignedTVList clone() {
-        return null;
-    }
+  @Override
+  public int compare(int idx1, int idx2) {
+    long t1 = getTime(idx1);
+    long t2 = getTime(idx2);
+    return Long.compare(t1, t2);
+  }
 
-    @Override
-    public int compare(int idx1, int idx2) {
-        long t1 = getTime(idx1);
-        long t2 = getTime(idx2);
-        return Long.compare(t1, t2);
-    }
-
-    @Override
-    public void swap(int p, int q) {
-        int valueP = getValueIndex(p);
-        long timeP = getTime(p);
-        int valueQ = getValueIndex(q);
-        long timeQ = getTime(q);
-        set(p, timeQ, valueQ);
-        set(q, timeP, valueP);
-    }
+  @Override
+  public void swap(int p, int q) {
+    int valueP = getValueIndex(p);
+    long timeP = getTime(p);
+    int valueQ = getValueIndex(q);
+    long timeQ = getTime(q);
+    set(p, timeQ, valueQ);
+    set(q, timeP, valueP);
+  }
 }
